pyproject.toml
[project]
name = "llms-and-tictactoe"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "accelerate>=1.13.0",
    "gymnasium>=1.2.3",
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pip>=26.0.1",
    "pyyaml>=6.0",
    "seaborn>=0.13.2",
    "torch>=2.11.0",
    "transformers>=5.5.4",
]
